Now after looking at static for a while I've (finally) noticed a pattern in the way min values are calculated. If you take a closer look at the following snippets you will notice a repetition:
min = Math.min(abs(memoized[?]), min);
min = Math.min(abs(memoized[?] - memoized[0]), min);
min = Math.min(abs(memoized[?] - memoized[1]), min);
min = Math.min(abs(memoized[?] - memoized[2]), min);

This is a linear scan though an array that is build 1 element / iteration. This is one of the Ns in our N^2. So what if there was a data structure that would take at most log(n) for insertions and log(n) for interval searches? Hmm.. That would be nice. A few lines of code later we have a basic implementation that fails the tests. That is Great! :)